requirements.txt
click==6.6
coverage==4.2
decorator==4.0.10
ecdsa==0.13
Fabric3==1.12.post1
Flask==0.11.1
inflection==0.2.0
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
-e git://github.com/CompareMyFare/jsonschema.git#egg=jsonschema
Markdown==2.4
MarkupSafe==0.23
pandocfilters==1.2
paramiko==1.17.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
py==1.4.31
pycrypto==2.6.1
Pygments==2.1.3
pytest==3.0.4
pytest-cov==2.4.0
pytest-flask==0.10.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
Werkzeug==0.11.11
